Split canThis from index.js
